1 november changes

+++ DeVeeraApp/Areas/Admin/Controllers/ImageController.cs
@@ -81,10 +81,14 @@ namespace DeVeeraApp.Areas.Admin.Controllers
                 {
                     var data = model.ToEntity<Image>();
                     var imageUrl = UploadToAWS(model.FileName);
+                    var spanishimageUrl = UploadToAWS(model.SpanishFileName);
                     data.ImageUrl = imageUrl.Result.ToString();
                     data.Key = model.FileName;
                     data.CreatedOn = DateTime.Now;
                     data.UpdatedOn = DateTime.Now;
+
+                    data.SpanishKey = model.SpanishFileName;
+                    data.SpanishImageUrl = spanishimageUrl.Result.ToString();
                     _imageMasterService.InsertImage(data);
                     _notificationService.SuccessNotification("Image url added successfully");
                     return RedirectToAction("List");
@@ -126,11 +130,20 @@ namespace DeVeeraApp.Areas.Admin.Controllers
                     if (model.FileName != null)
                     {
                         var url = UploadToAWS(model.FileName);
+                       
                         imageData.ImageUrl = url.Result;
+                        
                         imageData.Key = model.FileName;
-                        imageData.UpdatedOn = DateTime.Now;
-
+                        
                     }
+
+                    if (model.SpanishFileName != null)
+                    {
+                        var spanishurl = UploadToAWS(model.SpanishFileName);
+                        imageData.SpanishImageUrl = spanishurl.Result;
+                        imageData.SpanishKey = model.SpanishFileName;
+                    }
+                    imageData.UpdatedOn = DateTime.Now;
                 }
                 catch
                 { }
@@ -219,7 +232,7 @@ namespace DeVeeraApp.Areas.Admin.Controllers
             return Json(new { selectedImage = selectedImage, ImageFieldId = model.ImageFieldId, ImageFieldUrl = model.ImageFieldUrl });
         }
 
-        public IActionResult DeleteImage(int imageId)
+        public IActionResult DeleteImage(int imageId,int imagekeyval)
         {
             ResponseModel response = new ResponseModel();
 
@@ -232,11 +245,19 @@ namespace DeVeeraApp.Areas.Admin.Controllers
                     response.Success = false;
                     response.Message = "No video found";
                 }
+                if (imagekeyval == 1) { 
                 _s3BucketService.DeleteFile(data.Key);
 
                 data.Key = null;
                 data.ImageUrl = null;
+                }
+                if (imagekeyval == 2)
+                {
+                    _s3BucketService.DeleteFile(data.SpanishKey);
 
+                    data.SpanishKey = null;
+                    data.SpanishImageUrl = null;
+                }
                 _imageMasterService.UpdateImage(data);
                 response.Success = true;
